requirements-gpu.txt
Numpy==1.16.4
tensorflow_gpu==1.13.1
Keras==2.2.4
nltk==3.2.5
pillow==8.4.0
Matplotlib==3.3.1
tqdm==4.28.1
streamlit
googletrans
gtts

